+++ sisl/geom/basic.py
@@ -6,7 +6,7 @@ import numpy as np
 from sisl._internal import set_module
 from sisl import Geometry, SuperCell
 
-__all__ = ['sc', 'bcc', 'fcc', 'hcp']
+__all__ = ['sc', 'bcc', 'fcc', 'hcp', 'rocksalt']
 
 # A few needed variables
 _s30 = 1 / 2
@@ -37,6 +37,8 @@ def sc(alat, atom):
     g = Geometry([0, 0, 0], atom, sc=sc)
     if np.all(g.maxR(True) > 0.):
         g.optimize_nsc()
+    else:
+        g.set_nsc([3, 3, 3])
     return g
 
 
@@ -66,6 +68,8 @@ def bcc(alat, atoms, orthogonal=False):
         g = Geometry([0, 0, 0], atoms, sc=sc)
     if np.all(g.maxR(True) > 0.):
         g.optimize_nsc()
+    else:
+        g.set_nsc([3, 3, 3])
     return g
 
 
@@ -140,4 +144,39 @@ def hcp(a, atoms, coa=1.63333, orthogonal=False):
                      atoms, sc=sc)
     if np.all(g.maxR(True) > 0.):
         g.optimize_nsc()
+    else:
+        g.set_nsc([3, 3, 3])
+    return g
+
+
+@set_module("sisl.geom")
+def rocksalt(alat, atoms, orthogonal=False):
+    """ Two-element rocksalt lattice with 2 (non-orthogonal) or 8 atoms (orthogonal)
+
+    This is equivalent to the NaCl crystal structure (halite).
+
+    Parameters
+    ----------
+    alat : float
+        lattice parameter
+    atoms : list
+        a list of two atoms that the crystal consists of
+    orthogonal : bool, optional
+        whether the lattice is orthogonal or not
+    """
+    if isinstance(atoms, str):
+        atoms = [atoms, atoms]
+    if len(atoms) != 2:
+        raise ValueError(f"Invalid list of atoms, must have length 2")
+    g1 = fcc(alat, atoms[0], orthogonal=orthogonal)
+    g2 = fcc(alat, atoms[1], orthogonal=orthogonal).move(np.array([1, 1, 1]) * alat / 2)
+    g = g1.add(g2)
+    d = np.ones(3) * 1e-4
+    g = g.move(d).translate2uc().move(-d)
+    g.xyz = np.where(g.xyz > 0, g.xyz, 0)
+    g = g.sort(lattice=[2, 1, 0])
+    if np.all(g.maxR(True) > 0.):
+        g.optimize_nsc()
+    else:
+        g.set_nsc([3, 3, 3])
     return g